+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Using `extract_cluster_assignment()`, `extract_centroids()` and `predict()` on a fitted `hier_clust()` model without specifying `num_clust` or `cut_height` now gives more informative error message. (#147)
 
+* `k_means()` now errors informatively if `fit()` without `num_clust` specified. (#134)
+
 # tidyclust 0.1.2
 
 * The cluster specification methods for `generics::tune_args()` and `generics::tunable()` are now registered unconditionally (#115).